[KIE-792] introducing prototype dialect
