+++ drools-impact-analysis/drools-impact-analysis-parser/src/main/java/org/drools/impact/analysis/parser/impl/RhsParser.java
@@ -43,7 +43,7 @@ import org.drools.impact.analysis.model.right.ModifiedMapProperty;
 import org.drools.impact.analysis.model.right.ModifiedProperty;
 import org.drools.impact.analysis.model.right.ModifyAction;
 import org.drools.model.codegen.execmodel.generator.Consequence;
-import org.drools.model.codegen.execmodel.generator.DeclarationSpec;
+import org.drools.model.codegen.execmodel.generator.TypedDeclarationSpec;
 import org.drools.model.codegen.execmodel.generator.RuleContext;
 
 import static org.drools.util.StringUtils.ucFirst;
@@ -97,7 +97,7 @@ public class RhsParser {
         Expression actionArg = statement.getArgument(0);
         Class<?> actionClass = null;
         if (actionArg.isNameExpr()) {
-            actionClass = context.getDeclarationById(actionArg.toString()).map(DeclarationSpec::getDeclarationClass)
+            actionClass = context.getDeclarationById(actionArg.toString()).map(TypedDeclarationSpec::getDeclarationClass)
                     .orElseGet(() -> getClassFromAssignment(consequenceExpr, actionArg));
         } else if (actionArg.isLiteralExpr()) {
             actionClass = literalType(actionArg.asLiteralExpr());
