+++ drools-model/drools-model-codegen/src/main/java/org/drools/model/codegen/execmodel/generator/OOPathExprGenerator.java
@@ -137,7 +137,7 @@ public class OOPathExprGenerator {
             bindingId = context.getOOPathId(fieldType, originalBind + fieldName + pos);
         }
 
-        DeclarationSpec newDeclaration = context.addDeclaration(bindingId, fieldType, reactiveFrom);
+        TypedDeclarationSpec newDeclaration = context.addDeclaration(bindingId, fieldType, reactiveFrom);
         context.addOOPathDeclaration(newDeclaration);
 
         final List<DrlxExpression> conditions = chunk.getConditions();
