+++ drools-model/drools-model-codegen/src/main/java/org/drools/model/codegen/execmodel/generator/visitor/FromVisitor.java
@@ -37,7 +37,7 @@ import com.github.javaparser.ast.stmt.ExpressionStmt;
 import org.drools.drl.ast.descr.FromDescr;
 import org.drools.drl.ast.descr.PatternSourceDescr;
 import org.drools.model.codegen.execmodel.errors.InvalidExpressionErrorResult;
-import org.drools.model.codegen.execmodel.generator.DeclarationSpec;
+import org.drools.model.codegen.execmodel.generator.TypedDeclarationSpec;
 import org.drools.model.codegen.execmodel.generator.DrlxParseUtil;
 import org.drools.model.codegen.execmodel.generator.RuleContext;
 import org.drools.model.codegen.execmodel.generator.TypedExpression;
@@ -121,7 +121,7 @@ public class FromVisitor {
             if (isEnumeratedList( expr )) {
                 asListCall.addArgument( createListForLiteralFrom( expr.substring( 1, expr.length()-1 ), fromCall, usedDeclarations ) );
             } else {
-                Optional<DeclarationSpec> optContainsBinding = context.getDeclarationById( expr );
+                Optional<TypedDeclarationSpec> optContainsBinding = context.getDeclarationById(expr );
                 if ( optContainsBinding.isPresent() ) {
                     String bindingId = optContainsBinding.get().getBindingId();
                     fromCall.addArgument( context.getVarExpr( bindingId ) );
@@ -269,7 +269,7 @@ public class FromVisitor {
 
     private Expression createArg(String expression, String bindingId, MethodCallExpr fromCall) {
         if (bindingId != null) {
-            DeclarationSpec declarationSpec = context.getDeclarationById( bindingId ).orElseThrow( RuntimeException::new );
+            TypedDeclarationSpec declarationSpec = context.getDeclarationById(bindingId ).orElseThrow(RuntimeException::new );
             Class<?> clazz = declarationSpec.getDeclarationClass();
 
             DrlxParseResult drlxParseResult = ConstraintParser.withoutVariableValidationConstraintParser(context, context.getPackageModel())
