+++ drools-model/drools-model-codegen/src/main/java/org/drools/model/codegen/execmodel/generator/visitor/pattern/PatternAccumulateConstraint.java
@@ -68,7 +68,7 @@ class PatternAccumulateConstraint implements DSLNode {
 
         constraintsByVar.forEach( (id, constraints) -> {
             pattern.setIdentifier(id);
-            new PatternDSLPattern(context, packageModel, pattern, constraints, null).buildPattern();
+            new ClassPatternDSL(context, packageModel, pattern, constraints, null).buildPattern();
         });
         pattern.setIdentifier(null);
     }
