+++ drools-model/drools-model-codegen/src/main/java/org/drools/model/codegen/execmodel/generator/visitor/pattern/PatternConstraintParseResult.java
@@ -20,26 +20,4 @@ package org.drools.model.codegen.execmodel.generator.visitor.pattern;
 
 import org.drools.model.codegen.execmodel.generator.drlxparse.DrlxParseResult;
 
-public class PatternConstraintParseResult {
-    private String expression;
-    private String patternIdentifier;
-    private DrlxParseResult drlxParseResult;
-
-    public PatternConstraintParseResult(String expression, String patternIdentifier, DrlxParseResult drlxParseResult) {
-        this.expression = expression;
-        this.patternIdentifier = patternIdentifier;
-        this.drlxParseResult = drlxParseResult;
-    }
-
-    public String getExpression() {
-        return expression;
-    }
-
-    public String getPatternIdentifier() {
-        return patternIdentifier;
-    }
-
-    public DrlxParseResult getDrlxParseResult() {
-        return drlxParseResult;
-    }
-}
+public record PatternConstraintParseResult(String expression, String patternIdentifier, DrlxParseResult drlxParseResult) { }
